1. add early stop for lr & nn 2. change eval score from int to double to support raw score

+++ src/main/java/ml/shifu/shifu/actor/worker/ScoreModelWorker.java
@@ -126,7 +126,7 @@ public class ScoreModelWorker extends AbstractWorkerActor {
                 buf.append("|" + csResult.getMedianScore());
 
                 // score
-                for(Integer score: csResult.getScores()) {
+                for(Double score: csResult.getScores()) {
                     buf.append("|" + score);
                 }
